Refactor Whisper transcription handling: move the serialization context manager and transcription logic to tests.utils for improved modularity and clarity in conftest.py. Update audio file conversion to utilize the new subprocess function.
